+++ src/rust/build.rs
@@ -11,9 +11,6 @@ fn main() {
     if let Ok(version) = env::var("DEP_OPENSSL_VERSION_NUMBER") {
         let version = u64::from_str_radix(&version, 16).unwrap();
 
-        if version >= 0x3_00_00_00_0 {
-            println!("cargo:rustc-cfg=CRYPTOGRAPHY_OPENSSL_300_OR_GREATER");
-        }
         if version >= 0x3_00_09_00_0 {
             println!("cargo:rustc-cfg=CRYPTOGRAPHY_OPENSSL_309_OR_GREATER");
         }