+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Changelog
 
 .. note:: This version is not yet released and is under active development.
 
+* **BACKWARDS INCOMPATIBLE:** Support for OpenSSL 1.1.x has been removed.
+  OpenSSL 3.0.0 or later is now required. LibreSSL, BoringSSL, and AWS-LC
+  continue to be supported.
 * Support for ``x86_64`` macOS (including publishing wheels) is deprecated
   and will be removed in the next release. We will switch to publishing an
   ``arm64`` only wheel for macOS.